+++ api/transport/outbound.go
@@ -64,8 +64,16 @@ type OnewayOutbound interface {
 	CallOneway(ctx context.Context, request *Request) (Ack, error)
 }
 
-// Outbounds encapsulates outbound types for a service
+// Outbounds encapsulates the outbound specification for a service.
+//
+// This includes the service name that will be used for outbound requests
+// as well as the Outbound that will be used to transport the request.  The
+// outbound will be one of:
+//
+// - Unary (send request and wait for response)
+// - Oneway (send request and continue once downstream acknowledges request)
 type Outbounds struct {
-	Unary  UnaryOutbound
-	Oneway OnewayOutbound
+	ServiceName string
+	Unary       UnaryOutbound
+	Oneway      OnewayOutbound
 }